+++ rbx_types/src/content.rs
@@ -49,6 +49,30 @@ impl Content {
     pub fn value_mut(&mut self) -> &mut ContentType {
         &mut self.0
     }
+
+    /// Consumes this `Content` and returns the underlying value.
+    #[inline]
+    pub fn into_value(self) -> ContentType {
+        self.0
+    }
+
+    /// If this `Content` is a URI, returns the URI. Otherwise, returns `None`.
+    #[inline]
+    pub fn as_uri(&self) -> Option<&str> {
+        match self.value() {
+            ContentType::Uri(uri) => Some(uri),
+            _ => None,
+        }
+    }
+
+    /// If this `Content` is an Object, returns the Ref. Otherwise, returns `None`.
+    #[inline]
+    pub fn as_object(&self) -> Option<Ref> {
+        match self.value() {
+            &ContentType::Object(referent) => Some(referent),
+            _ => None,
+        }
+    }
 }
 
 impl From<String> for Content {
@@ -72,7 +96,6 @@ impl From<&'_ str> for Content {
     derive(serde::Serialize, serde::Deserialize),
     serde(transparent)
 )]
-
 pub struct ContentId {
     url: String,
 }